Fixes for compiling with VC6.
--HG--
branch : xqilla

+++ src/yajl/yajl_gen.h
@@ -62,7 +62,7 @@ extern "C" {
     void YAJL_API yajl_gen_free(yajl_gen handle);
 
     yajl_gen_status YAJL_API yajl_gen_integer(yajl_gen hand,
-                                              long long int number);
+                                              long int number);
     yajl_gen_status YAJL_API yajl_gen_double(yajl_gen hand, double number);
     yajl_gen_status YAJL_API yajl_gen_string(yajl_gen hand,
                                              const unsigned char * str,
